whatpm/t/tree-test-phrasing.dat

#data
</sarcasm>
#errors
no DOCTYPE
unmatched end tag
#document
| <html>
|   <head>
|   <body>

#data
<sarcasm>aaa</sarcasm>
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <sarcasm>
|       "aaa"
